Some work to thread everything through - creating the pledge, end to end, pending confirmation page

+++ src/SFA.DAS.LevyTransferMatching.Infrastructure/Dto/AccountDto.cs
@@ -6,6 +6,6 @@ namespace SFA.DAS.LevyTransferMatching.Infrastructure.Dto
 {
     public class AccountDto
     {
-        public int RemainingTransferAllowance { get; set; }
+        public double RemainingTransferAllowance { get; set; }
     }
 }
